refactor: directions and geocode
