+++ package.json
@@ -21,7 +21,6 @@
     "@tanstack/react-query": "^5.66.0",
     "@tanstack/react-query-persist-client": "^5.101.4",
     "ajv": "^8.18.0",
-    "axios": "^1.16.1",
     "babel-plugin-react-compiler": "^1.0.0",
     "base64-arraybuffer": "^1.0.2",
     "expo": "^56",
